backend/requirements.txt
flask
minio
python-dotenv
kafka-python-ng
flask-cors